feat: Refactor SchemaControler to make it stateless - fixed curcular dependencies
