+++ src/Adapters/Schema/SchemaCacheAdapter.js
@@ -2,24 +2,16 @@
 /**
  * @module Adapters
  */
-import type { Schema } from '../../Controllers/types';
-import { SchemaData } from '../../Controllers/SchemaController';
+import { SchemaAndData } from './types';
 
 /**
  * @interface SchemaCacheAdapter
  */
 export default class SchemaCacheAdapter {
   /**
-   * Used by controller to provide data from database
-   */
-  setDataProvider(
-    dataProvider: () => Promise<{ allClasses: Array<Schema>, schemaData: SchemaData }>
-  ) {}
-
-  /**
    * Get all schema entries and its corresponding intermediate format
    */
-  async fetchSchema(): Promise<{ allClasses: Array<Schema>, schemaData: SchemaData }> {}
+  async fetchSchema(getDataFromDb: () => Promise<SchemaAndData>): Promise<SchemaAndData> {}
 
   /**
    * Clear cache
